remove gasPrice.test.ts as we do not use it and move the keystore test to the main tests directory
